Fixed the insert function for members. You have to specify the schema and make sure you put the correct name of the table.
Still have to fix the update function for members, but I think it's not working because I had put the wrong table AND because the query has memberIdInput at the end and that should be reflected in the other files (members.jsp, DataHandler.java). I'm tired, so I'll do this tomorrow.